requirements.txt
numpy>=1.24
Pillow>=9.5